fix: fix variable type

+++ src/main/java/com/part3/team07/sb01deokhugamteam07/dto/book/PopularBookDto.java
@@ -1,6 +1,7 @@
 package com.part3.team07.sb01deokhugamteam07.dto.book;
 
 import com.part3.team07.sb01deokhugamteam07.entity.Period;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.UUID;
 
@@ -12,9 +13,9 @@ public record PopularBookDto(
     String thumbnailFileName,
     Period period,
     int rank,
-    double score,
+    BigDecimal score,
     int reviewCount,
-    double rating,
+    BigDecimal rating,
     LocalDateTime createdAt
 ) {
 }
